Change program counter to use 16 bit int

+++ src/vm/slip_vm.h
@@ -95,7 +95,7 @@ struct SlipVM {
     SlipByte keys[SLIP_KEYS];
 
     // The program counter.
-    size_t PC;
+    int16_t PC;
 
 };
